In epd analysis consider -eval option value.
If -eval static, then the ce value of the epd should be taken from
engine static eval.

if -eval search, then the ce value of the epd output should be taken
from the engine search score.

If -eval none, then no analysis will be done.

If -eval is not specified, then static will be used as this is the
default value of -eval in epd analysis.